wa-gateway/requirements.txt
fastapi[standard]
python-dotenv
httpx
pika